# pqarrow: use ColumnIndex to answer distinct queries at the scan level