This commit uses a column chunk's column index to optimize distinct queries
where there is a single value in the chunk (across all pages) of a
dictionary-encoded column. In this case, a dictionary can be constructed with
the value stored in the index, avoiding an expensive page decompression step.

```
name          old time/op    new time/op    delta
QueryTypes-8     169ms ±11%     133ms ±18%  -21.14%  (p=0.049 n=10+3)

name          old alloc/op   new alloc/op   delta
QueryTypes-8     443MB ± 0%     436MB ± 0%   -1.54%  (p=0.007 n=10+3)

name          old allocs/op  new allocs/op  delta
QueryTypes-8     1.13M ± 0%     1.05M ± 0%   -6.78%  (p=0.007 n=10+3)
```

Closes #158 

cc @metalmatze @thorfour