@asubiotto 2022-08-18
I took a look at what's failing and it's a type assertion in parquet when I attempt to create a dictionary from a string in the new code path. The underlying `columnType` that I call `NewDictionary` on is a `string` type wrapped by `indexedType`, but `indexedType` overloads its wrapped type's `NewValues` method with a call to `Int32ValuesFromBytes`, which means that the values passed in to `NewDictionary` are int32s. I'm guessing that these are the indexes of the indexedType so not a bug (which indicates to me that I need to create a dictionary in another way), but this behavior seems weird.