@asubiotto 2022-08-22
Thanks all! I would say to skip the failing test, merge, and unskip once the upstream bug is fixed, but it looks like there are a couple more tests failing due to the same reason. I'll wait to merge this until https://github.com/segmentio/parquet-go/issues/312 is fixed.