@asubiotto 2022-08-26
Found a way to bypass the parquet panic (using a passed in explicit type instead of determining it from the column chunk). Will merge on Monday.